rfc: rfcBBL209 stay strictly trustless for now.
Rather than build off #25, I think it better to do all the parts prior
to #25, so we have less complexity to wrap our heads around when
evaluating the security properties etc.

Generalizing as done with #25 I agree is a good idea, but can always be
done as a follow-up step.